Update 063-mapping-rf.Rmd
Hi, kindly see line 241 where pred variable is being back-transformed with the exp function. 

I see that on line 174 above, you applied the back-transformation to the pred variable and it is the same pred variable being referenced on line 241 which makes the back-transformation redundant on line 241 (in my opinion...Please correct me if I am wrong). 

I think the exp function should be removed from line 241. Also, when exp is applied on line 241, the pred values for OCSKGM go out of the range of the original values.